client/requirements.txt
python-telegram-bot==13.15
vosk==0.3.45
SQLAlchemy==2.0.20
pydub==0.25.1
python-dotenv==1.0.0
requests==2.31.0
